Prioritize the hostname for determining environment-specific constants.
The backend development setup runs a proxy to the production instance of the
community site, and expects the community site to detect that it is running
under the glitch.development hostname and make API requests to
api.glitch.development. Prior to this fix, the RUNNING_ON environment varible
overrode this hostname detection and made requests to the production API even
when running under the glitch.development hostname.

This updates the existing logic so that the RUNNING_ON environment variable
doesn't override the hostname detection that the backend dev setup expects.